Move debug.h to public/oscap_debug.h
Also update includes and Makefile appropriately.
