+++ src/common/Makefile.am
@@ -4,7 +4,7 @@ liboscapcommon_la_SOURCES = \
 	alloc.c alloc.h \
 	assume.h \
 	bfind.c bfind.h \
-	debug.c debug.h debug_priv.h \
+	debug.c debug_priv.h \
 	elements.c elements.h \
 	err_queue.c err_queue.h \
 	error.c _error.h \
@@ -35,6 +35,7 @@ liboscapcommon_la_LIBADD = \
 	@xml2_LIBS@ @xslt_LIBS@ @exslt_LIBS@ @pthread_LIBS@
 
 pkginclude_HEADERS =\
+	public/oscap_debug.h \
 	public/oscap_error.h \
 	public/oscap.h \
 	public/oscap_reference.h \
